{"schema":"decodex/commit/1","summary":"Modularize state store runtime code","authority":"XY-1115"}
